# docs/swagger.yaml
basePath: /
definitions:
  github_com_awang-karisma_trustpilot-scraper_internal_api_dto.CreateWebsiteRequest:
    properties:
      base_url:
        type: string
      enabled:
        type: boolean
      name:
        type: string
      schedule:
        type: string
    required:
    - base_url
    - name
    - schedule
    type: object
  github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  github_com_awang-karisma_trustpilot-scraper_internal_api_dto.SuccessResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  github_com_awang-karisma_trustpilot-scraper_internal_api_dto.UpdateWebsiteRequest:
    properties:
      base_url:
        type: string
      enabled:
        type: boolean
      name:
        minLength: 1
        type: string
      schedule:
        type: string
    type: object
  github_com_awang-karisma_trustpilot-scraper_internal_api_dto.WebhookTestRequest:
    properties:
      template_path:
        type: string
      webhook_type:
        type: string
      webhook_url:
        type: string
    required:
    - webhook_url
    type: object
  github_com_awang-karisma_trustpilot-scraper_internal_api_dto.WebhookTestResponse:
    properties:
      message:
        type: string
      response_status:
        type: integer
      success:
        type: boolean
    type: object
  github_com_awang-karisma_trustpilot-scraper_internal_api_dto.WebhookTriggerRequest:
    properties:
      review_id:
        description: 'Optional: specific review to send'
        type: integer
    type: object
  github_com_awang-karisma_trustpilot-scraper_internal_api_dto.WebhookTriggerResponse:
    properties:
      message:
        type: string
      sent_at:
        type: string
      success:
        type: boolean
    type: object
  github_com_awang-karisma_trustpilot-scraper_internal_api_dto.WebsiteListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.WebsiteResponse'
        type: array
      total:
        type: integer
    type: object
  github_com_awang-karisma_trustpilot-scraper_internal_api_dto.WebsiteResponse:
    properties:
      base_url:
        type: string
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: integer
      last_scraped:
        type: string
      name:
        type: string
      schedule:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: This is a Trustpilot Scraper API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Trustpilot Scraper API
  version: "1.0"
paths:
...
  /api/webhook/test:
    post:
      consumes:
      - application/json
      description: Test a webhook configuration with mock data
      parameters:
      - description: Webhook test configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.WebhookTestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.WebhookTestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse'
      summary: Test a webhook
      tags:
      - webhooks
  /api/websites:
    get:
      description: Get a list of all registered websites to scrape
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.WebsiteListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse'
      summary: List all websites
      tags:
      - websites
    post:
      consumes:
      - application/json
      description: Register a new website for scraping
      parameters:
      - description: Website details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.CreateWebsiteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.WebsiteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse'
      summary: Create a new website
      tags:
      - websites
  /api/websites/{id}:
    delete:
      description: Delete a registered website by ID
      parameters:
      - description: Website ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse'
      summary: Delete a website
      tags:
      - websites
    get:
      description: Get a registered website by ID
      parameters:
      - description: Website ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.WebsiteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse'
      summary: Get a website
      tags:
      - websites
    put:
      consumes:
      - application/json
      description: Update registered website details by ID
      parameters:
      - description: Website ID
        in: path
        name: id
        required: true
        type: integer
      - description: Website details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.UpdateWebsiteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.WebsiteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse'
      summary: Update a website
      tags:
      - websites
...
  /api/websites/{id}/scrape:
    post:
      description: Manually trigger an immediate scraping job for a website
      parameters:
      - description: Website ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse'
      summary: Trigger a scrape
      tags:
      - websites
  /api/websites/{id}/webhook:
    post:
      consumes:
      - application/json
      description: Manually trigger a webhook for a specific website
      parameters:
      - description: Website ID
        in: path
        name: id
        required: true
        type: integer
      - description: Webhook trigger options
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.WebhookTriggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.WebhookTriggerResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_awang-karisma_trustpilot-scraper_internal_api_dto.ErrorResponse'
      summary: Trigger a webhook
      tags:
      - webhooks
swagger: "2.0"